+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.11.0] - 2024-12-05
+### Fixed
+
+- Disallow 0 from being passed for disk_iops and update the docs to indicate
+  that omitting the attribute is the correct way to get the default behavior.
 
 ### Added