+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Disallow 0 from being passed for disk_iops and update the docs to indicate
+- Update docs to reflect that maintenance windows are for Advanced clusters
+  only.
+
+- Disallow 0 from being passed for `disk_iops` and update the docs to indicate
   that omitting the attribute is the correct way to get the default behavior.
 
 - Validate that `node_count` is not passed in with serverless cluster regions.