+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.3] - 2025-02-27
+
+### Fixed
+
+- Remove UseStateForUnknown plan modifier for state parameter in cluster resource.
+
 ## [1.11.2] - 2025-02-07
 
 ### Fixed